Add modulator packet size tag handling
The modulator does have a certain input output ratio that would confuse
packet size tags. This problem is fixed now. It looks for tags that
carry packet size and deletes them. Afterwards it adds new tags with the
appropriate packet size.

Also, this commit adds in 'extract_burst' to make use of frequency
offset estimation and may compensate if activated.