Add short burst shaper
This commit adds 'Short Burst Shaper' to the module. It is required to
transmit non-continuous burst with e.g. a USRP. Essentially it is a zero
pre-pad/post-pad block wrapped in a `tagged_stream_block`. Otherwise,
burst stubs end up in the USRP sink and result in partially transmitted
bursts.